fix ops.pad bf16 with fp32 cast

+++ examples/opensora_hpcai/opensora/models/vae/vae_temporal.py
@@ -462,9 +462,9 @@ class VAE_Temporal(nn.Cell):
         )
         # x = pad_at_dim(x, (time_padding, 0), dim=2)
 
-        # NOTE: due to ops.pad not support bf16
         x_dtype = x.dtype
-        x = x.to(ms.float16)
+        # FIXME: bf16 not supported for ops.pad. float32 can be slow but ensure no loss. TODO: check speed difference with fp16
+        x = x.to(ms.float32)
         x = ops.pad(x, (0, 0, 0, 0, time_padding, 0), mode="constant")
         x = x.to(x_dtype)
